FIXED:   pbx/monitoring:  Audio Files about meetme was not delete if they do not exist. Other Problem about delete files if one of then don't exist the folder monitor was removed by using exec("rm -rf $dir")

+++ modules/branches/1.6/core/monitoring/index.php
@@ -79,7 +79,7 @@ function _moduleContent(&$smarty, $module_name)
 
     switch($action){
         case 'delete':
-            $content = deleteRecord($smarty, $module_name, $local_templates_dir, $pDB, $pDBACL, $arrConf);
+            $content = deleteRecord($smarty, $module_name, $local_templates_dir, $pDB, $pDBACL, $arrConf, $arrLang);
             break;
         case 'download':
             $content = downloadFile($smarty, $module_name, $local_templates_dir, $pDB, $pDBACL, $arrConf);
@@ -350,7 +350,7 @@ contenido;
         $smarty->display("_common/popup.tpl");
 }
 
-function deleteRecord($smarty, $module_name, $local_templates_dir, &$pDB, &$pDBACL, $arrConf)
+function deleteRecord($smarty, $module_name, $local_templates_dir, &$pDB, &$pDBACL, $arrConf, $arrLang)
 {
     $pMonitoring = new paloSantoMonitoring($pDB);
     $_DATA['date_start'] = isset($date_ini)?$date_ini:date("d M Y");
@@ -359,16 +359,19 @@ function deleteRecord($smarty, $module_name, $local_templates_dir, &$pDB, &$pDBA
     foreach($_POST as $key => $values){
         if(substr($key,0,3) == "id_")
         {
-            $ID = substr($key, 3);
+			$ID = substr($key, 3);
             $ID = str_replace("_",".",$ID);
             $recordName = $pMonitoring->getRecordName($ID);
-
-            if($pMonitoring->deleteRecordFile($ID)){
-                $record = substr($recordName,6);
-                $path = $path_record.$record;
-                exec("rm -rf $path");
-            }
-        }
+            $record = substr($recordName,6);
+            $record = basename($record);
+            $path = $path_record.$record;
+            if(is_file($path)){
+                if($pMonitoring->deleteRecordFile($ID))
+                    unlink($path);
+            }else{
+                $smarty->assign("mb_message", $arrLang['delete_file_error']);
+            }        
+         }
     }
     $_POST = $_DATA;
     $content = reportMonitoring($smarty, $module_name, $local_templates_dir, $pDB, $pDBACL, $arrConf);
